refact: abstract ContributionTimeline and expose edit activity
