+++ components/contributions/ContributionDrawer.tsx
@@ -1,43 +1,20 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { gql, useQuery } from '@apollo/client';
-import clsx from 'clsx';
-import dayjs from 'dayjs';
-import { groupBy, isEmpty, sortBy, uniq } from 'lodash';
-import {
-  Calendar,
-  Check,
-  ChevronDown,
-  ChevronUp,
-  CircleCheckBig,
-  CircleStop,
-  Hourglass,
-  Info,
-  Link as LinkIcon,
-  MoreHorizontalIcon,
-  Plus,
-  RefreshCcw,
-  TriangleAlert,
-  Undo2,
-  Wallet,
-  X,
-} from 'lucide-react';
+import { isEmpty } from 'lodash';
+import { Check, Link as LinkIcon, X } from 'lucide-react';
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 
 import type { GetActions } from '../../lib/actions/types';
 import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
 import type { ContributionDrawerQuery, ContributionDrawerQueryVariables } from '../../lib/graphql/types/v2/graphql';
-import { ActivityType, ContributionFrequency, OrderStatus, TransactionKind } from '../../lib/graphql/types/v2/graphql';
+import { ContributionFrequency, OrderStatus } from '../../lib/graphql/types/v2/graphql';
 import useClipboard from '../../lib/hooks/useClipboard';
 import useLoggedInUser from '../../lib/hooks/useLoggedInUser';
 import { i18nPaymentMethodProviderType } from '../../lib/i18n/payment-method-provider-type';
 import type LoggedInUser from '../../lib/LoggedInUser';
-import { getDashboardRoute } from '../../lib/url-helpers';
 
 import { AccountHoverCard } from '../AccountHoverCard';
 import Avatar from '../Avatar';
-import ActivityDescription from '../dashboard/sections/ActivityLog/ActivityDescription';
-import { useTransactionActions } from '../dashboard/sections/transactions/actions';
-import DateTime from '../DateTime';
 import FormattedMoneyAmount from '../FormattedMoneyAmount';
 import Link from '../Link';
 import LoadingPlaceholder from '../LoadingPlaceholder';
@@ -48,15 +25,11 @@ import PaymentMethodTypeWithIcon from '../PaymentMethodTypeWithIcon';
 import { DropdownActionItem } from '../table/RowActionsMenu';
 import Tags from '../Tags';
 import { Button } from '../ui/Button';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '../ui/DropdownMenu';
+import { DropdownMenu, DropdownMenuContent, DropdownMenuSeparator, DropdownMenuTrigger } from '../ui/DropdownMenu';
 import { Sheet, SheetContent, SheetFooter } from '../ui/Sheet';
 import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/Tooltip';
+
+import ContributionTimeline, { getTransactionsUrl } from './ContributionTimeline';
 
 type ContributionDrawerProps = {
   open: boolean;
@@ -79,6 +52,12 @@ const I18nFrequencyMessages = defineMessages({
     defaultMessage: 'Yearly',
   },
 });
+
+function getTransactionOrderLink(LoggedInUser: LoggedInUser, order: ContributionDrawerQuery['order']): string {
+  const url = getTransactionsUrl(LoggedInUser, order);
+  url.searchParams.set('orderId', order.legacyId.toString());
+  return url.toString();
+}
 
 export function ContributionDrawer(props: ContributionDrawerProps) {
   const clipboard = useClipboard();
@@ -116,6 +95,9 @@ export function ContributionDrawer(props: ContributionDrawerProps) {
             id
             name
             description
+          }
+          createdByAccount {
+            ...ContributionDrawerAccountFields
           }
           fromAccount {
             ...ContributionDrawerAccountFields
@@ -561,7 +543,7 @@ export function ContributionDrawer(props: ContributionDrawerProps) {
                       <LoadingPlaceholder height={20} />
                     </div>
                   ) : (
-                    <OrderTimeline order={query.data.order} />
+                    <ContributionTimeline order={query.data.order} />
                   )}
                 </div>
               </div>
@@ -588,371 +570,5 @@ export function ContributionDrawer(props: ContributionDrawerProps) {
         </SheetFooter>
       </SheetContent>
     </Sheet>
-  );
-}
-
-type TransactionGroup = {
-  group: string;
-  transactions: ContributionDrawerQuery['order']['transactions'];
-};
-
-function buildTransactionGroups(order: ContributionDrawerQuery['order']): TransactionGroup[] {
-  const transactions = order?.transactions ?? [];
-  if (transactions.length === 0) {
-    return [];
-  }
-
-  const groups = uniq(transactions.map(txn => txn.group));
-  const byGroup = groupBy(transactions, txn => txn.group);
-
-  const KIND_PRIORITY = [TransactionKind.CONTRIBUTION, TransactionKind.ADDED_FUNDS, TransactionKind.BALANCE_TRANSFER];
-  return groups.map(group => ({
-    group,
-    transactions: byGroup[group].sort((a, b) => {
-      if (KIND_PRIORITY.includes(a.kind) && !KIND_PRIORITY.includes(b.kind)) {
-        return -1;
-      } else if (!KIND_PRIORITY.includes(a.kind) && KIND_PRIORITY.includes(b.kind)) {
-        return 1;
-      }
-      return a.createdAt - b.createdAt;
-    }),
-  }));
-}
-
-function getTransactionsUrl(LoggedInUser: LoggedInUser, order: ContributionDrawerQuery['order']): URL {
-  let route = `/${order.toAccount.slug}/transactions`;
-  if (LoggedInUser.isSelf(order.fromAccount)) {
-    route = getDashboardRoute(order.fromAccount, 'transactions');
-  } else if ('host' in order.toAccount && LoggedInUser.canSeeAdminPanel(order.toAccount.host)) {
-    route = getDashboardRoute(order.toAccount.host, 'host-transactions');
-  } else if (LoggedInUser.canSeeAdminPanel(order.toAccount)) {
-    route = getDashboardRoute(order.toAccount, 'transactions');
-  }
-
-  return new URL(route, window.location.origin);
-}
-
-function getTransactionGroupLink(
-  LoggedInUser: LoggedInUser,
-  order: ContributionDrawerQuery['order'],
-  transactionGroup: string,
-): string {
-  const url = getTransactionsUrl(LoggedInUser, order);
-  url.searchParams.set('group', transactionGroup);
-  return url.toString();
-}
-
-function getTransactionOrderLink(LoggedInUser: LoggedInUser, order: ContributionDrawerQuery['order']): string {
-  const url = getTransactionsUrl(LoggedInUser, order);
-  url.searchParams.set('orderId', order.legacyId.toString());
-  return url.toString();
-}
-
-type OrderTimelineProps = {
-  order: ContributionDrawerQuery['order'];
-};
-
-type OrderTimelineItem = {
-  id: string;
-  title: React.ReactNode;
-  icon: React.ReactNode;
-  date: string | Date;
-  expected?: boolean;
-  type?: 'error' | 'warning' | 'success' | 'info';
-  collapsable?: boolean;
-  collapseGroup?: string;
-  normallyOpen?: boolean;
-  menu?: React.ReactNode;
-};
-
-function OrderTimeline(props: OrderTimelineProps) {
-  const { LoggedInUser } = useLoggedInUser();
-  const [collapseGroupsToggle, setCollapseGroupsToggle] = useState({});
-  const getTransactionActions = useTransactionActions();
-
-  const toggleGroup = React.useCallback((group: string) => {
-    setCollapseGroupsToggle(cur => {
-      if (Object.hasOwn(cur, group)) {
-        const newValue = { ...cur };
-        delete newValue[group];
-        return newValue;
-      } else {
-        return { ...cur, [group]: true };
-      }
-    });
-  }, []);
-
-  const activities: OrderTimelineItem[] = props.order.activities.nodes
-    .filter(a => {
-      if (
-        [
-          ActivityType.COLLECTIVE_TRANSACTION_CREATED,
-          ActivityType.ORDER_PENDING_CREATED,
-          ActivityType.ORDER_PENDING_RECEIVED,
-          ActivityType.ORDER_PENDING,
-        ].includes(a.type)
-      ) {
-        return false;
-      }
-
-      return true;
-    })
-    .map(a => {
-      return {
-        id: a.id,
-        title: <ActivityDescription activity={a} />,
-        collapsable: a.type === ActivityType.PAYMENT_FAILED,
-        icon: [
-          ActivityType.PAYMENT_CREDITCARD_CONFIRMATION,
-          ActivityType.ORDER_PAYMENT_FAILED,
-          ActivityType.PAYMENT_FAILED,
-          ActivityType.ORDER_PENDING_EXPIRED,
-          ActivityType.ORDER_DISPUTE_CREATED,
-          ActivityType.ORDER_DISPUTE_CLOSED,
-          ActivityType.CONTRIBUTION_REJECTED,
-        ].includes(a.type) ? (
-          <TriangleAlert size={16} />
-        ) : a.type === ActivityType.SUBSCRIPTION_CANCELED ? (
-          <CircleStop size={16} />
-        ) : [ActivityType.ORDER_PROCESSING, ActivityType.ORDER_PENDING_CONTRIBUTION_REMINDER].includes(a.type) ? (
-          <Hourglass size={16} />
-        ) : (
-          <Info size={16} />
-        ),
-        date: a.createdAt,
-        type: [
-          // TODO: add disputed, in review
-          ActivityType.ORDER_PAYMENT_FAILED,
-          ActivityType.SUBSCRIPTION_CANCELED,
-          ActivityType.PAYMENT_FAILED,
-          ActivityType.ORDER_PENDING_EXPIRED,
-          ActivityType.CONTRIBUTION_REJECTED,
-        ].includes(a.type)
-          ? 'error'
-          : [
-                ActivityType.PAYMENT_CREDITCARD_CONFIRMATION,
-                ActivityType.ORDER_PENDING_CONTRIBUTION_REMINDER,
-                ActivityType.ORDER_DISPUTE_CREATED,
-                ActivityType.ORDER_DISPUTE_CLOSED,
-              ].includes(a.type)
-            ? 'warning'
-            : 'info',
-        menu: a.data?.paymentProcessorUrl ? (
-          <React.Fragment>
-            <DropdownMenuItem asChild>
-              <Button size="xs" variant="ghost" asChild className="w-full">
-                <Link href={a.data?.paymentProcessorUrl} className="justify-start">
-                  <FormattedMessage defaultMessage="View in payment processor" id="NgSLbI" />
-                </Link>
-              </Button>
-            </DropdownMenuItem>
-          </React.Fragment>
-        ) : null,
-      };
-    });
-
-  const transactions: OrderTimelineItem[] = sortBy(
-    buildTransactionGroups(props.order),
-    txn => txn.transactions.at(0).createdAt,
-  )
-    .reverse()
-    .map((txn, i) => {
-      const primaryTxn = txn.transactions.at(0);
-      const actions = getTransactionActions(primaryTxn);
-
-      return {
-        id: txn.group,
-        title: primaryTxn.description,
-        icon: primaryTxn.isRefund ? <Undo2 size={16} /> : <CircleCheckBig size={16} />,
-        date: primaryTxn.createdAt,
-        collapsable: true,
-        type: primaryTxn.isRefund ? 'warning' : 'success',
-        normallyOpen: i === 0,
-        menu: (
-          <React.Fragment>
-            <DropdownMenuItem asChild>
-              <Button size="xs" variant="ghost" asChild className="w-full justify-start">
-                <Link href={getTransactionGroupLink(LoggedInUser, props.order, txn.group)}>
-                  <FormattedMessage defaultMessage="View transaction" id="1kZ3H0" />
-                </Link>
-              </Button>
-            </DropdownMenuItem>
-            {primaryTxn.paymentProcessorUrl && (
-              <DropdownMenuItem asChild>
-                <Button size="xs" variant="ghost" asChild className="w-full justify-start">
-                  <Link href={primaryTxn.paymentProcessorUrl}>
-                    <FormattedMessage defaultMessage="View in payment processor" id="NgSLbI" />
-                  </Link>
-                </Button>
-              </DropdownMenuItem>
-            )}
-            {actions.primary?.map((action, i) => (
-              // eslint-disable-next-line react/no-array-index-key
-              <DropdownMenuItem key={i} asChild className="w-full">
-                <Button size="xs" variant="ghost" onClick={action.onClick} className="justify-start">
-                  {action.label}
-                </Button>
-              </DropdownMenuItem>
-            ))}
-          </React.Fragment>
-        ),
-      };
-    });
-
-  const otherActivities: OrderTimelineItem[] = [];
-
-  if (
-    props.order.frequency !== ContributionFrequency.ONETIME &&
-    [OrderStatus.ACTIVE, OrderStatus.PROCESSING].includes(props.order.status) &&
-    dayjs(props.order.nextChargeDate).isAfter(new Date())
-  ) {
-    otherActivities.push({
-      id: 'next-charge-date',
-      title: <FormattedMessage defaultMessage="Next subscription cycle" id="qzsea6" />,
-      icon: <Wallet size={16} />,
-      date: props.order.nextChargeDate as string,
-      expected: true,
-      type: 'info',
-    });
-  }
-
-  const latestProcessingEvent = props.order.activities.nodes.findLast(a => a.type === ActivityType.ORDER_PROCESSING);
-  const latestChargeEvent = transactions.length > 0 ? sortBy(transactions, 'date').at(-1) : null;
-
-  if (
-    props.order.status === OrderStatus.PROCESSING &&
-    latestProcessingEvent &&
-    (!latestChargeEvent || dayjs(latestChargeEvent.date).isBefore(dayjs(latestProcessingEvent.createdAt)))
-  ) {
-    // is waiting processing.
-    const expectedChargeDate = dayjs(latestProcessingEvent.createdAt).add(14, 'day');
-    otherActivities.push({
-      id: 'expected-processing-end',
-      title: <FormattedMessage defaultMessage="Expected date to complete processing charge" id="26Lp/C" />,
-      icon: <Calendar size={16} />,
-      date: expectedChargeDate.toISOString(),
-      expected: true,
-      type: 'info',
-    });
-  }
-
-  if (
-    props.order.frequency !== ContributionFrequency.ONETIME &&
-    props.order.status === OrderStatus.ERROR &&
-    dayjs(props.order.nextChargeDate).isAfter(new Date())
-  ) {
-    otherActivities.push({
-      id: 'next-charge-attempt',
-      title: <FormattedMessage defaultMessage="Next charge attempt" id="ud3Qe6" />,
-      icon: <RefreshCcw size={16} />,
-      date: props.order.nextChargeDate as string,
-      expected: true,
-      type: 'info',
-    });
-  }
-
-  if (props.order.status === OrderStatus.PENDING && props.order.pendingContributionData?.expectedAt) {
-    otherActivities.push({
-      id: 'expected-funds',
-      title: <FormattedMessage defaultMessage="Expected date for funds received" id="2f9kCJ" />,
-      icon: <Calendar size={16} />,
-      date: props.order.pendingContributionData.expectedAt as string,
-      expected: true,
-      type: 'info',
-    });
-  }
-
-  otherActivities.push({
-    id: 'order-created',
-    title: <FormattedMessage defaultMessage="Order created" id="XNf7EA" />,
-    icon: <Plus size={16} />,
-    date: props.order.createdAt as string,
-    collapsable: true,
-    type: 'info',
-  });
-
-  let lastCollapseGroup;
-  const timeline: OrderTimelineItem[] = sortBy([...activities, ...transactions, ...otherActivities], 'date')
-    .reverse()
-    .map((item, i, arr) => {
-      const lastItem = i - 1 >= 0 ? arr[i - 1] : null;
-
-      if (lastItem?.collapsable && !item.collapsable) {
-        lastCollapseGroup = lastItem.id;
-        return { ...item, collapseGroup: lastItem.id };
-      } else if (lastCollapseGroup && !item.collapsable) {
-        return { ...item, collapseGroup: lastCollapseGroup };
-      } else {
-        lastCollapseGroup = null;
-      }
-
-      return item;
-    });
-
-  return (
-    <ol className="mt-3 pl-3">
-      {timeline.map((t, itemIndex) => {
-        const collapseParent = t.collapseGroup ? timeline.find(item => item.id === t.collapseGroup) : null;
-        if (
-          collapseParent &&
-          (collapseParent.normallyOpen
-            ? Object.hasOwn(collapseGroupsToggle, collapseParent.id)
-            : !Object.hasOwn(collapseGroupsToggle, collapseParent.id))
-        ) {
-          return null;
-        }
-        const hasChild = t.collapsable ? timeline.some(item => item.collapseGroup === t.id) : false;
-        const isCollapsed =
-          hasChild &&
-          t.collapsable &&
-          (t.normallyOpen ? Object.hasOwn(collapseGroupsToggle, t.id) : !Object.hasOwn(collapseGroupsToggle, t.id));
-        return (
-          <React.Fragment key={t.id}>
-            <li
-              className={clsx('relative border-l border-[#E1E7EF] pb-10 pl-9 last:border-none last:pb-0', {
-                'border-dashed': t.expected,
-              })}
-            >
-              <div
-                className={clsx('absolute -left-3 flex h-6 w-6 items-center justify-center rounded-full ring-8', {
-                  'bg-[#FFF0F1] text-[#FF3053] ring-[#FFF0F1]': t.type === 'error',
-                  'bg-[#FEFEE8] text-[#A27006] ring-[#FEFEE8]': t.type === 'warning',
-                  'bg-[#EAFFE9] text-[#098605] ring-[#EAFFE9]': t.type === 'success',
-                  'bg-[#EEF7FF] text-[#308EFF] ring-[#EEF7FF]': t.type === 'info',
-                })}
-              >
-                {t.icon}
-              </div>
-              <div className="relative flex items-center">
-                {itemIndex !== 0 && <div className="absolute -top-5 w-full border-t border-[#E1E7EF]"></div>}
-                <div>
-                  <div>{t.title}</div>
-                  <div className="text-xs leading-5 text-[#75777A]">
-                    <DateTime value={t.date} timeStyle="short" />
-                  </div>
-                </div>
-                <div className="ml-auto flex gap-2">
-                  {t.menu && (
-                    <DropdownMenu>
-                      <DropdownMenuTrigger asChild>
-                        <Button variant="ghost" size="icon-xs">
-                          <MoreHorizontalIcon size={16} />
-                        </Button>
-                      </DropdownMenuTrigger>
-                      <DropdownMenuContent>{t.menu}</DropdownMenuContent>
-                    </DropdownMenu>
-                  )}
-                  {t.collapsable && hasChild && (
-                    <Button variant="ghost" size="icon-xs" onClick={() => toggleGroup(t.id)}>
-                      {isCollapsed ? <ChevronDown size={16} /> : <ChevronUp size={16} />}
-                    </Button>
-                  )}
-                </div>
-              </div>
-            </li>
-          </React.Fragment>
-        );
-      })}
-    </ol>
   );
 }
