+++ lib/i18n/activities.ts
@@ -540,6 +540,10 @@ export const ActivityDescriptionI18n = defineMessages({
     defaultMessage: 'Vendor <Vendor></Vendor> created',
     id: 'IUlgDG',
   },
+  ADDED_FUNDS_EDITED: {
+    defaultMessage: '<Individual></Individual> edited added funds',
+    id: 'nxtsKq',
+  },
 });
 
 type ActivityTranslations = Partial<Record<keyof typeof ActivityTypes, MessageDescriptor>>;
